add DI rabbit Recevie
