+++ WorkerServiceEmail/WorkerServiceEmail/Properties/launchSettings.json
@@ -6,8 +6,8 @@
       "environmentVariables": {
         "DOTNET_ENVIRONMENT": "Development",
         "LOG_DIRECTORY": "D:/NewTemp",
-        "LOGIN_EMAIL_GMAIL": "dogsitterclub2022@gmail.com",
-        "PASSWORD_EMAIL_GMAIL": "devedu2022!",
+        "LOGIN_EMAIL_GMAIL": "",
+        "PASSWORD_EMAIL_GMAIL": "",
         "LOGIN_EMAIL_YANDEX": "",
         "PASSWORD_EMAIL_YANDEX": ""
       }
