+++ WorkerServiceEmail/WorkerServiceEmail/Services/Interfaces/IRabbitReceiveService.cs
@@ -1,0 +1,7 @@
+﻿namespace WorkerServiceEmail.Services
+{
+    public interface IRabbitReceiveService
+    {
+        void Recevie();
+    }
+}
